+++ lib/mainScreen.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:material_symbols_icons/symbols.dart';
 import 'package:sedong_mbti/const/sizes.dart';
-import 'package:sedong_mbti/view/devel_mbti.dart';
+import 'package:sedong_mbti/mbtiScreen/devel_mbti.dart';
 import 'package:sedong_mbti/view/roulette.dart';
 
 class mainScreen extends StatefulWidget {
