+++ lib/mbtiScreen/Screen2.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class Screen2 extends StatefulWidget {
+  const Screen2({super.key});
+
+  @override
+  State<Screen2> createState() => _Screen2State();
+}
+
+class _Screen2State extends State<Screen2> {
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold();
+  }
+}
