+++ lib/mbtiScreen/mbti_Button/QustionsAnswer.dart
@@ -1,0 +1,31 @@
+// import 'package:flutter/material.dart';
+// import 'package:sedong_mbti/const/sizes.dart';
+
+// import 'package:sedong_mbti/mbtiScreen/mbti_Button/QustionButton.dart';
+
+// class QustionsAnwser extends StatefulWidget {
+//   String qus;
+//   String answer1;
+//   String answer2;
+//   QustionsAnwser({
+//     super.key,
+//     required this.qus,
+//     required this.answer1,
+//     required this.answer2,
+//   });
+
+//   @override
+//   State<QustionsAnwser> createState() => _QustionsAnwserState();
+// }
+
+// class _QustionsAnwserState extends State<QustionsAnwser> {
+//   @override
+//   Widget build(BuildContext context) {
+//     return Column(
+//       mainAxisAlignment: MainAxisAlignment.center,
+//       children: [
+        
+//       ],
+//     );
+//   }
+// }
