+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   kbspinningwheel: ^1.0.3
   material_symbols_icons: ^4.2719.1
   font_awesome_flutter: ^10.7.0
+  page_transition: ^2.1.0
